rgw: Approach to fix uninitialised RGWCopyObj executed as admin.
This is a proposal of change in order to verify that a given rgw
operation is fully initialised before bypassing possible errors when
checking for an operation when running it as admin user.

For example, running the
`s3tests_boto3.functional.test_s3:test_object_copy_to_itself` s3 test
as admin user ends up crashing the gateway.

This adds a new method (`check_initialised`) to the `RGWOp` interface
that is called before bypassing any possible error detected when
checking for permissions.

If the op is not fully initialised (which means this error was detected
before the initialisation) it simply returns the error and avoids
executing the op (which will much likely end up in undesired results).

This is just a proposal to open a debate in conjuntion with opening the bug and only adds
checks for the `RGWCopyObj` operation.

Fixes: https://tracker.ceph.com/issues/58035

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>