rgw/sfs: Lifecycle initial implementation
Initial PR implementing lifecycle.
This PR tries to use the existing rgw's lifecycle, so all the
lifecycle functinalities that upstream `rgw` support should be also
supported (obviously if the features needed are already implemented in
`sfs`).

It also takes benefit of sharding which is done by rgw's code.

When a bucket is changed to add lifecycle confuguration it is added to a
shard id based on a hash function.

Every shard has a `head` which stores the point where a particular shard
id is being processed.

The advantage of splitting things into buckets is being able to run
lifecycle checks and actions in parallel tasks.

Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>