rgw/sfs: GC remove deleted objects and nocurrent versions
Implements permanent deletion of objects that have a delete marker, plus
permenent deletion of noncurrent versions.

Versions to be deleted (may imply deleting the whole object) are
obtained from a query to the metadata to delete them by size in
descending order.

In order to trigger GC on demand I think we can do it by simply calling:
`store->gc->process()` as the garbage collection instance is public
within the `SFStore`.

That will simply run the same process that it's executed by the garbage
collector internal timer.

Fixes: https://github.com/aquarist-labs/s3gw/issues/393
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>