rgw/sfs: Delete markers on top
This PR changes how we query for last version, and how delete markers
are obtained.

It sorts by version_type first, so commit time is only taken into
account if a delete marker is not found.
It also selects with a limit(1).

It also refactors a few other version queries to make them simpler and
to make them one of them not a READ-WRITE transaction.

Fixes: https://github.com/aquarist-labs/s3gw/issues/622
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>