rgw/sfs: use sqlite3 backup API to create temp db
It's not safe for DBConn::check_metadata_is_compatible()
to make a copy of the database file while it's already open
via SQLite, because the close() that happens after the file
is copied cancels all POSIX advisory locks, which can later
lead to database corruption.  For details, see:

https://www.sqlite.org/howtocorrupt.html#_posix_advisory_locks_canceled_by_a_separate_thread_doing_close_

This change replaces fs::copy() with use of SQLite's Backup API
(https://www.sqlite.org/backup.html) which is safe to use with
open databases.

Fixes: https://github.com/aquarist-labs/s3gw/issues/702
Signed-off-by: Tim Serong <tserong@suse.com>