rgw/sfs: testing: destroy dbconn and store before removing db dir
TestSFSConcurrency and TestSFSList both keep pointers to a
sqlite::DBConnRef and a rgw::sal::SFStore.  The TearDown() method
deletes the database directory while these objects still exist,
and then later when they're automatically destroyed, we see the
following output when running the unit tests:

[SQLITE] (28) file unlinked while open: /tmp/8kGFAz3iWQxhP9SR6oTflAF/s3gw.db
[SQLITE] (28) file unlinked while open: /tmp/8kGFAz3iWQxhP9SR6oTflAF/s3gw.db

This happens many times (twice for each test for each of those
classes).  If we reset the pointers before removing the database
directory, the objects are destroyed in advance and the errors go
away.

Signed-off-by: Tim Serong <tserong@suse.com>