rgw/sfs: Fix GC gets stuck when deleteing a bucket
This fixes a bug found in GC after running the multipart integration
tests.

The issue was caused by multiparts that had no parts, that were not
considered when deleting the multiparts of a deleted bucket.

Before permanently removing a bucket GC cancels all the pending
multiparts and then removes all multiparts but multiparts with 0 parts
were not removed.

GC tried to delete the bucket and, as there were multiparts still in the
database, the delete bucket transaction failed.

GC deletes stuff in batches, which means that things are deleted in
loops, we should also take into consideration what to do when a
transaction fails forever in a loop.

I leave that part for a future PR after discussing with the team.

Fixes: https://github.com/aquarist-labs/s3gw/issues/719
Signed-off-by: Xavi Garcia <xavi.garcia@suse.com>